Validador + test

+++ src/main/java/Models/FormasDeContribucion/DonacionDeDinero.java
@@ -8,10 +8,10 @@ import java.time.LocalDate;
 @Getter
 public class DonacionDeDinero extends FormaDeContribucion {
     private LocalDate fechaDeDonacion;
-    private Integer monto;
+    private Double monto;
     private TipoFrecuencia frecuencia;
 
-    public DonacionDeDinero(Integer monto, TipoFrecuencia frecuencia){
+    public DonacionDeDinero(Double monto, TipoFrecuencia frecuencia){
         this.monto = monto;
         this.frecuencia = frecuencia;
         this.fechaDeDonacion = LocalDate.now();
